Refactor: Interactor 를 View 기준으로 할당 - View 가 할 수 있는 Interaction 을 명확하게 만들기 위함
