+++ Umpa/Presentation/Main/Community/Components/HotPostsBanner.swift
