+++ Umpa/Presentation/Main/Home/Components/TeacherFindingCarouselItem.swift
