+++ Umpa/Presentation/Main/Matching/ServiceDetail/MrCreationService/MrCreationServiceDetailView.swift
@@ -5,17 +5,17 @@ import Domain
 import Factory
 import SwiftUI
 
-struct LessonServiceDetailView: ServiceDetailView {
+struct MrCreationServiceDetailView: ServiceDetailView {
     @InjectedObject(\.appState) private var appState
     @Injected(\.chatInteractor) private var chatInteractor
-    @Injected(\.serviceInteractor) private var serviceInteractor
+    @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
 
-    let service: LessonService
+    let service: MusicCreationService
 
     var tabItems: [TabItem] {
-        service.curriculum.isEmpty
-            ? [.teacherOverview, .lessonOverview, .review]
-            : [.teacherOverview, .lessonOverview, .curriculum, .review]
+        service.sampleMusics.isEmpty
+            ? [.teacherOverview, .serviceOverview, .review]
+            : [.teacherOverview, .serviceOverview, .samplePreview, .review]
     }
 
     @State private var tabSelection = 0
@@ -28,7 +28,6 @@ struct LessonServiceDetailView: ServiceDetailView {
             }
     }
 
-    @ViewBuilder
     var content: some View {
         ZStack(alignment: .bottom) {
             ScrollView {
@@ -44,7 +43,7 @@ struct LessonServiceDetailView: ServiceDetailView {
                 height: bottomActionBarHeight,
                 isLiked: false, // TODO: isLiked 를 받아와야 함
                 likeButtonAction: { isLiked in
-                    serviceInteractor.markAsLike(isLiked, for: service.id)
+                    serviceDetailInteractor.markAsLike(isLiked, for: service.id)
                 },
                 primaryButtonAction: {
                     chatInteractor.startChatting(
@@ -61,31 +60,31 @@ struct LessonServiceDetailView: ServiceDetailView {
         switch tabItems[tabSelection] {
         case .teacherOverview:
             TeacherOverviewTabContent(teacher: service.author)
-        case .lessonOverview:
-            LessonOverviewTabContent(service: service)
-        case .curriculum:
-            CurriculumTabContent(curriculumList: service.curriculum)
+        case .serviceOverview:
+            ServiceOverviewTabContent(service: service)
+        case .samplePreview:
+            SamplePreviewTabContent(sampleMusics: service.sampleMusics)
         case .review:
             ReviewTabContent(service: service)
         }
     }
 }
 
-extension LessonServiceDetailView {
+extension MrCreationServiceDetailView {
     enum TabItem {
         case teacherOverview
-        case lessonOverview
-        case curriculum
+        case serviceOverview
+        case samplePreview
         case review
 
         var name: String {
             switch self {
             case .teacherOverview:
                 return "선생님 소개"
-            case .lessonOverview:
-                return "수업 소개"
-            case .curriculum:
-                return "커리큘럼"
+            case .serviceOverview:
+                return "서비스 안내"
+            case .samplePreview:
+                return "샘플 확인"
             case .review:
                 return "리뷰"
             }
@@ -95,12 +94,6 @@ extension LessonServiceDetailView {
 
 #if MOCK
 #Preview {
-    @Injected(\.appState) var appState
-    appState.userData.currentUser = Student.sample0
-
-    return
-        NavigationStack {
-            LessonServiceDetailView(service: .sample0)
-        }
+    MrCreationServiceDetailView(service: .sample0)
 }
 #endif
